Change RETURNS slightly and introduce RETURNS_DECAYED
Ignore-this: 253e64058f523f69692b4180ad7fa503
This is a breaking change: RETURNS now returns the type of the expression, not a variable.
This will not usually make a difference but it might in same cases.
If the decayed type is required, RETURNS_DECAYED is now provided.

darcs-hash:20150408091743-c9a90-c3c6b3da58f74d58c59406148371473ba0518315